semester-3/mlops/assignment1/mlops-heart-disease/requirements.txt
# -----------------------------------
# Model-critical (EXACT / COMPATIBLE)
# -----------------------------------
scikit-learn==1.7.2
numpy>=1.23,<=2.3.5
joblib==1.5.2
# -------------------------------
# Data & experiment tracking
# -------------------------------
pandas>=2.3,<3.0
mlflow>=3.7,<4.0
python-dotenv>=1.0

# -------------------------------
# API runtime
# -------------------------------
fastapi>=0.124
uvicorn>=0.38

# -------------------------------
# Testing
# -------------------------------
pytest>=9.0

# --------------------------------------
# Added for visualization and reporting
# --------------------------------------
matplotlib>=3.9,<4.0
seaborn>=0.13,<0.14
prometheus-client